requirements.txt
discord
python-dotenv
langchain
langchain-openai
langchain-tavily